Add useServiceWorker hook
