+++ src/App.js
@@ -1,21 +1,19 @@
 // @flow
 import { isMobile } from 'mobile-device-detect'
-import React, { memo, useCallback, useEffect, useState } from 'react'
+import React, { memo, useCallback, useState } from 'react'
 import { Platform, SafeAreaView, StyleSheet } from 'react-native'
 import PaperProvider from 'react-native-paper/src/core/Provider'
 import './lib/gundb/gundb'
+import useServiceWorker from './lib/hooks/useServiceWorker'
 import { theme } from './components/theme/styles'
-import SimpleStore, { setInitFunctions } from './lib/undux/SimpleStore'
+import SimpleStore from './lib/undux/SimpleStore'
 import RouterSelector from './RouterSelector.web'
 import LoadingIndicator from './components/common/view/LoadingIndicator'
 import SplashDesktop from './components/splash/SplashDesktop'
-import isWebApp from './lib/utils/isWebApp'
 import logger from './lib/logger/pino-logger'
 import { SimpleStoreDialog } from './components/common/dialogs/CustomDialog'
 import Config from './config/config'
-import * as serviceWorker from './serviceWorker'
 const log = logger.child({ from: 'App' })
-let serviceWorkerRegistred = false
 
 const SplashOrRouter = memo(({ store }) => {
   const isLoggedIn = !!store.get('isLoggedIn')
@@ -30,39 +28,9 @@ const SplashOrRouter = memo(({ store }) => {
 })
 
 const App = () => {
+  useServiceWorker()
+  log.debug({ Config })
   const store = SimpleStore.useStore()
-  useEffect(() => {
-    const onUpdate = reg => {
-      store.set('serviceWorkerUpdated')(reg)
-      navigator.serviceWorker.addEventListener('controllerchange', function() {
-        log.debug('service worker: controllerchange')
-        window.location.reload()
-      })
-    }
-    const onRegister = reg => {
-      //force check for service worker update
-      reg.update()
-      if (reg.waiting) {
-        onUpdate(reg)
-      }
-    }
-    if (serviceWorkerRegistred === false) {
-      log.debug('registering service worker')
-      serviceWorker.register({ onRegister, onUpdate })
-      serviceWorkerRegistred = true
-    }
-    if (isWebApp === false) {
-      log.debug('useEffect, registering beforeinstallprompt')
-
-      window.addEventListener('beforeinstallprompt', e => {
-        // For older browsers
-        e.preventDefault()
-        log.debug('Install Prompt fired')
-        store.set('installPrompt')(e)
-      })
-    }
-    setInitFunctions(store.set('wallet'), store.set('userStorage'))
-  }, [])
 
   // onRecaptcha = (token: string) => {
   //   userStorage.setProfileField('recaptcha', token, 'private')
